\name{nprobust-internal}
\alias{nprobust-internal}
\alias{W.fun}
\alias{h.cer.den}
\alias{hrbcpp}
\alias{k.fun}
\alias{kern.fun}
\alias{m.fun}
\alias{v.fun}
\alias{qrXXinv}
\alias{lprobust_vce}
\alias{lprobust_bw}
\alias{lprobust_res}
\title{Package Internal Functions}
\description{
Internal functions which are not part of the package application programming interface. 
}
\author{
Sebastian Calonico, University of Miami, Coral Gables, FL.  \email{scalonico@bus.miami.edu}.

Matias D. Cattaneo, University of Michigan, Ann Arbor, MI.  \email{cattaneo@umich.edu}.

Max H. Farrell, University of Chicago, Chicago, IL. \email{max.farrell@chicagobooth.edu}.
}

