\name{nprobust-package}
\alias{nprobust-package}
\alias{nprobust}
\title{Robust Data-Driven Statistical Inference for Local Polynomial Regression and Kernel Density Estimation}
\description{
   This package provides tools for data-driven  analytical statistical inference for Local Polynomial Regression (LPR) estimators and Kernel Density Estimation (KDE): \code{\link{lprobust}} to construct local-polynomial point estimators and robust confidence intervals, \code{\link{lpbwselect}} to perform bandwidth selection for the different procedures implemented.}
\details{
\tabular{ll}{
Package: \tab nprobust\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2016-11-06\cr
License: \tab GPL-2\cr
}
Function for LPR statistical inference: \code{\link{lprobust}}\cr
Function for LPR bandwidths selection: \code{\link{lpbwselect}}\cr
Function for KDE statistical inference: \code{\link{kdrobust}}\cr
Function for KDE bandwidths selection: \code{\link{kdbwselect}}\cr
}
\author{
Sebastian Calonico, University of Miami, Coral Gables, FL.  \email{scalonico@bus.miami.edu}.

Matias D. Cattaneo, University of Michigan, Ann Arbor, MI.  \email{cattaneo@umich.edu}.

Max H. Farrell, University of Chicago, Chicago, IL. \email{max.farrell@chicagobooth.edu}.
}

