\name{nprobust-package}
\alias{nprobust-package}
\alias{nprobust}
\title{Nonparametric Robust Estimation and Inference Methods using Local Polynomial Regression and Kernel Density Estimation}
\description{
   This package provides tools for data-driven statistical analysis using local polynomial regression (LPR) and kernel density estimation (KDE) methods: \code{\link{lprobust}} for local polynomial point estimation and robust bias-corrected inference, \code{\link{lpbwselect}} for local polynomial bandwidth selection, \code{\link{kdrobust}} for kernel density point estimation and robust bias-corrected inference, \code{\link{kdbwselect}} for kernel density bandwidth selection, and \code{\link{nprobust.plot}} for plotting results.}
\details{
\tabular{ll}{
Package: \tab nprobust\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2017-09-14\cr
License: \tab GPL-2\cr 
}
Function for LPR estimation and inference: \code{\link{lprobust}}\cr
Function for LPR bandwidth selection: \code{\link{lpbwselect}}\cr
Function for KDE estimation and inference: \code{\link{kdrobust}}\cr
Function for KDE bandwidth selection: \code{\link{kdbwselect}}\cr
Function for graphical analysis: \code{\link{nprobust.plot}}\cr
}
\author{
Sebastian Calonico, University of Miami, Coral Gables, FL.  \email{scalonico@bus.miami.edu}.

Matias D. Cattaneo, University of Michigan, Ann Arbor, MI.  \email{cattaneo@umich.edu}.

Max H. Farrell, University of Chicago, Chicago, IL. \email{max.farrell@chicagobooth.edu}.
}

